+++ packages/rich-text-with-backend/src/components/RichText.tsx
@@ -2,9 +2,9 @@ import { Placeholder } from '@tiptap/extension-placeholder';
 import { EditorContent, JSONContent } from '@tiptap/react';
 import React, { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useBlockEditor } from './hooks/useBlockEditor';
-import i18n from './locales/localization';
-import './styles/index.css';
+import { useBlockEditor } from '../hooks/useBlockEditor';
+import i18n from '../locales/localization';
+import '../styles/index.css';
 
 type Props = {
   initialContent?: string;
